% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw_class.R
\docType{class}
\name{ddhw-class}
\alias{ddhw-class}
\title{An S4 class to represent the DDHW output.}
\description{
An S4 class to represent the DDHW output.
}
\section{Slots}{

\describe{
\item{\code{df}}{A data frame which collects the input data such as the vector of p values and the filter statistics, the group assignment, as well as outputs (weighted p-values, adjusted p-values)}

\item{\code{weights}}{A numeric vector of the weight assigned to each stratum}

\item{\code{thresholds}}{A numeric vector of the rejection threshold for each stratum}

\item{\code{alpha}}{Numeric, the nominal significance level at which the FDR is to be controlled}

\item{\code{regularization_term}}{Numeric, the value of the regularization parameter after the cross-val choice}

\item{\code{penalty}}{Character, "uniform deviation" or "total variation"}

\item{\code{reg_path_information}}{Information about the whole regularization path. !NOT implemented yet}

\item{\code{solver_information}}{Solver specific output, e.g. was the problem solved to optimality?}
}}
\seealso{
\code{\link{weights.ddhw}}, \code{\link{thresholds.ddhw}}, \code{\link{rejections.ddhw}}, \code{\link{adj_pvalues.ddhw}}
}

