% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw.R
\name{ddhw_auto}
\alias{ddhw_auto}
\title{ddhw_auto: Main function to do Data-Driven Hypothesis Weights}
\usage{
ddhw_auto(unadj_p, filterstat, alpha, nbins = "auto", reg_par_number = 7,
  reg_pars = "linear", test_distrib = "grenander", nreps = 10,
  sd_penalty = 0, verbosity_level = 0, ...)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{filterstat}{Vector which contains the filter-statistic (covariate, independent under the H0) for each p-value. This can be a numeric vector or a factor.}

\item{alpha}{Numeric, sets the nominal level for FDR control.}

\item{nbins}{Integer, number of groups into which p-values will be split based on filter_statistic (default = "auto").}

\item{reg_par_number}{Integer, number of regularization parameters to try out (length of grid)}

\item{reg_pars}{"linear" if sequence of regularization parameters should form a grid, "power" if it should be geometrically decreasing, a numeric vector if the user wants to explicitly prespecify the regularization parameter values to test}

\item{test_distrib}{Method to estimate the distribution function in the test set (after holdout split). Available "grenander" (Strimmer's modified Grenander estimator in fdrtools), "ecdf" (the ECDF), "both" (the maximum of "ecdf" and "grenander")}

\item{nreps}{Number of replications of the 2-fold CV procedure (i.e. 2*nreps total holdout runs)}

\item{sd_penalty}{Additional penalization before accepting a regularization parameter value}

\item{verbosity_level}{Verbosity level}
}
\value{
A ddhw object
}
\description{
ddhw_auto: Main function to do Data-Driven Hypothesis Weights
}
\seealso{
the parameters passed to ddhw / ddhw_grouped \code{\link{ddhw}}
}

