% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{groups_by_filter_multivariate}
\alias{groups_by_filter_multivariate}
\title{Stratify hypotheses based on increasing value of the covariate}
\usage{
groups_by_filter_multivariate(
  covariates,
  nbins,
  ties.method = "random",
  seed = NULL
)
}
\arguments{
\item{covariates}{matrix, data.frame or tibble of ordinal covariates based on which the stratification will be done.}

\item{nbins}{Integer, number of groups/strata into which p-values will be split based on covariate.}

\item{ties.method}{Character specifying how ties are treated, see \code{\link{rank}} function.}

\item{seed}{Integer, specifies random seed to be used when \code{ties.method=="random"}.}
}
\value{
A factor with nbins different levels, each entry corresponds to the stratum the i-th hypothesis
 was assigned to.
}
\description{
Hypotheses are stratified into nbins different strata of (approximately) equal size based on
increasing value of the covariate
extension of \code{\link{groups_by_filter}} for multidimensional covariates, per-covariate quantiles
}
\examples{
covariates <- matrix(runif(300), ncol = 3)
groups <- groups_by_filter_multivariate(covariates, 10)
}
