% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihw_class.R
\docType{class}
\name{ihwResult-class}
\alias{adj_pvalues}
\alias{adj_pvalues,ihwResult-method}
\alias{alpha}
\alias{alpha,ihwResult-method}
\alias{as.data.frame,ihwResult-method}
\alias{as.data.frame.ihwResult}
\alias{covariates}
\alias{covariates,ihwResult-method}
\alias{groups_factor}
\alias{groups_factor,ihwResult-method}
\alias{ihwResult}
\alias{ihwResult-class}
\alias{pvalues}
\alias{pvalues,ihwResult-method}
\alias{rejected_hypotheses}
\alias{rejected_hypotheses,ihwResult-method}
\alias{rejections}
\alias{rejections,ihwResult-method}
\alias{show,ihwResult-method}
\alias{thresholds}
\alias{thresholds,ihwResult-method}
\alias{weighted_pvalues}
\alias{weighted_pvalues,ihwResult-method}
\alias{weights,ihwResult-method}
\title{An S4 class to represent the ihw output.}
\usage{
adj_pvalues(object)

\S4method{adj_pvalues}{ihwResult}(object)

\S4method{weights}{ihwResult}(object, levels_only = FALSE)

thresholds(object, ...)

\S4method{thresholds}{ihwResult}(object, levels_only = FALSE)

pvalues(object)

\S4method{pvalues}{ihwResult}(object)

weighted_pvalues(object)

\S4method{weighted_pvalues}{ihwResult}(object)

covariates(object)

\S4method{covariates}{ihwResult}(object)

groups_factor(object)

\S4method{groups_factor}{ihwResult}(object)

alpha(object)

\S4method{alpha}{ihwResult}(object)

rejections(object, ...)

\S4method{rejections}{ihwResult}(object)

rejected_hypotheses(object, ...)

\S4method{rejected_hypotheses}{ihwResult}(object)

\method{as.data.frame}{ihwResult}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{as.data.frame}{ihwResult}(x, row.names = NULL, optional = FALSE,
  ...)

\S4method{show}{ihwResult}(object)
}
\arguments{
\item{object, x}{A ihwResult object as returned by a call to ihw(...)}

\item{levels_only}{Logical, if FALSE, return a vector of weights (thresholds) with one weight
(threshold) for each hypothesis, otherwise return a nfolds x nbins matrix of weights (thresholds)}

\item{...}{Parameters passed in to individual methods}
}
\description{
An S4 class to represent the ihw output.
}
\section{Methods (by generic)}{
\itemize{
\item \code{adj_pvalues}: Extract adjusted pvalues

\item \code{weights}: Extract weights

\item \code{thresholds}: Calculate ihw thresholds

\item \code{pvalues}: Extract pvalues

\item \code{weighted_pvalues}: Extract weighted pvalues

\item \code{covariates}: Extract filter statistics

\item \code{groups_factor}: Extract factor of stratification (grouping) variable

\item \code{alpha}: Extract nominal significance (alpha) level

\item \code{rejections}: Total number of rejected hypotheses by ihw procedure

\item \code{rejected_hypotheses}: Get a boolean vector of the rejected hypotheses

\item \code{as.data.frame}: Coerce ihwResult to data frame

\item \code{show}: Convenience method to show ihwResult object
}}
\section{Slots}{

\describe{
\item{\code{df}}{A data.frame that collects the input data, including the vector of p values and the covariate, the group assignment, as well as outputs (weighted p-values, adjusted p-values)}

\item{\code{weights}}{A (nbins X nfolds) matrix of the weight assigned to each stratum}

\item{\code{alpha}}{Numeric, the nominal significance level at which the FDR is to be controlled}

\item{\code{nbins}}{Integer, number of distinct levels into which the hypotheses were stratified}

\item{\code{nfolds}}{Integer, number of folds for pre-validation procedure}

\item{\code{regularization_term}}{Numeric vector, the final value of the regularization parameter within each fold}

\item{\code{penalty}}{Character, "uniform deviation" or "total variation"}

\item{\code{covariate_type}}{Character, "ordinal" or "nominal"}

\item{\code{reg_path_information}}{A data.frame, information about the whole regularization path. (Currently not used, thus empty)}

\item{\code{solver_information}}{A list, solver specific output, e.g. were all subproblems solved to optimality? (Currently empty list)}
}}
\examples{

save.seed <- .Random.seed; set.seed(1)
X   <- runif(n = 20000, min = 0.5, max = 4.5)       # Covariate
H   <- rbinom(n = length(X), size = 1, prob = 0.1)  # Is the null hypothesis (mean=0) true or false ?
Z   <- rnorm(n = length(X), mean = H * X)           # Z-score
.Random.seed <- save.seed

pvalue <- 1 - pnorm(Z)                              # pvalue
ihw_res <- ihw(pvalue, covariates, alpha = 0.1)
rejections(ihw_res)
colnames(as.data.frame(ihw_res))

}
\seealso{
ihw, plot_ihw
}

