% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plot,ihwResult,ANY-method}
\alias{plot,ihwResult,ANY-method}
\title{Convenience function to plot weights learned by ihw method}
\usage{
\S4method{plot}{ihwResult,ANY}(x, x_axis = "group",
  scale = covariate_type(x))
}
\arguments{
\item{x}{ihwResult Object}

\item{x_axis}{Character, currently only "group" supported.}

\item{scale}{Character, "ordinal" or "nominal"}
}
\value{
A ggplot2 object.
}
\description{
Convenience function to plot weights learned by ihw method
}
\examples{

   save.seed <- .Random.seed; set.seed(1)
   X   <- runif(20000, min=0.5, max=4.5) #covariate
   H   <- rbinom(20000,1,0.1)            #hypothesis true or false
   Z   <- rnorm(20000, H*X)              #Z-score
   .Random.seed <- save.seed
   pvalue <- 1-pnorm(Z)                  #pvalue
   ihw_res <- ihw(pvalue, X, .1)
   plot(ihw_res)

}

