% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plot_ddhw}
\alias{plot_ddhw}
\alias{plot_ddhw,ddhwResult-method}
\title{Convenience function to plot weights learned by DDHW method}
\usage{
plot_ddhw(x, ...)

\S4method{plot_ddhw}{ddhwResult}(x, x_axis = "group",
  scale = x@filter_statistic_type)
}
\arguments{
\item{x}{ddhwResult Object}

\item{...}{additional arguments passed on to individual methods}

\item{x_axis}{Character, currently only "group" supported.}

\item{scale}{Character, "ordinal" or "nominal"}
}
\value{
A ggplot2 object.
}
\description{
Convenience function to plot weights learned by DDHW method
}
\section{Methods (by class)}{
\itemize{
\item \code{ddhwResult}: 
}}
\examples{
set.seed(1)
   X   <- runif(20000, min=0.5, max=4.5) #covariate
   H   <- rbinom(20000,1,0.1)            #hypothesis true or false
   Z   <- rnorm(20000, H*X)              #Z-score
   pvalue <- 1-pnorm(Z)                  #pvalue
   ddhw_res <- ddhw(pvalue, X, .1)
	  plot_ddhw(ddhw_res)
}

