% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plot_ihw}
\alias{plot_ihw}
\alias{plot_ihw,ihwResult-method}
\title{Convenience function to plot weights learned by ihw method}
\usage{
plot_ihw(x, ...)

\S4method{plot_ihw}{ihwResult}(x, x_axis = "group",
  scale = x@covariate_type)
}
\arguments{
\item{x}{ihwResult Object}

\item{...}{additional arguments passed on to individual methods}

\item{x_axis}{Character, currently only "group" supported.}

\item{scale}{Character, "ordinal" or "nominal"}
}
\value{
A ggplot2 object.
}
\description{
Convenience function to plot weights learned by ihw method
}
\section{Methods (by class)}{
\itemize{
\item \code{ihwResult}: Further arguments
}}
\examples{

   save.seed <- .Random.seed; set.seed(1)
   X   <- runif(20000, min=0.5, max=4.5) #covariate
   H   <- rbinom(20000,1,0.1)            #hypothesis true or false
   Z   <- rnorm(20000, H*X)              #Z-score
   .Random.seed <- save.seed
   pvalue <- 1-pnorm(Z)                  #pvalue
   ihw_res <- ihw(pvalue, X, .1)
   plot_ihw(ihw_res)

}

