% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_weights_quantile_2d}
\alias{plot_weights_quantile_2d}
\title{2-dimensional quantile plotting}
\usage{
plot_weights_quantile_2d(ihw_quantiles, m_eval = 100, fold = 1)
}
\arguments{
\item{ihw_quantiles}{Object of class \code{ihwResult}}

\item{m_eval}{number of evaluation points on grid}

\item{fold}{Fold to be plotted.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot weights of ihw for quantile stratification method in 2-dimensional case.
}
\examples{

   save.seed <- .Random.seed; set.seed(1)
   X   <- runif(20000, min = 0.5, max = 4.5) # covariate
   H   <- rbinom(20000, 1, 0.1)              # hypothesis true or false
   Z   <- rnorm(20000, H*X)                  # z-score
   .Random.seed <- save.seed
   pvalue <- 1-pnorm(Z)                      # pvalue
   X   <- cbind(X, runif(20000))             # add noise
   ihw_res <- ihw(pvalue, X, .1)
   plot(ihw_res)
   
}
