% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_weights}
\alias{random_weights}
\title{Generate random weights.}
\usage{
random_weights(npars, N = 1)
}
\arguments{
\item{npars}{Sampling will be from (npars-1)-simplex (embedded in R^npars)}

\item{N}{Number of random weight combinations to be generated}
}
\description{
this is equivalent to sampling uniformly from the n-simplex
see: http://geomblog.blogspot.de/2005/10/sampling-from-simplex.html
}

