% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weights.R
\name{regularize_thresholds}
\alias{regularize_thresholds}
\title{Fix numerically unstable MILP thresholds (Helper function)}
\usage{
regularize_thresholds(ts, m_groups, rjs, alpha, lambda,
  penalty = "total variation", solver = "Rsymphony")
}
\arguments{
\item{ts}{Thresholds returned from MILP (corresponding to highest rejected p-value in each group)}

\item{m_groups}{Number of tests in each stratum}

\item{rjs}{Number of rejections at MILP solver optimum}

\item{alpha}{Nominal level of multiple testing procedure}

\item{lambda}{Regularization parameter (so that total_variation(weights) <= lambda)}
}
\value{
Thresholds >= ts, which correspond to weights obeying the regularization condition
}
\description{
second small linear program to be solved if MILP was numerically unstable
then we try to manually check if total variation regularization can really be
 fullfilled
}

