% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw_class.R
\name{rejected_hypotheses}
\alias{rejected_hypotheses}
\title{Generic function for multiple testing procedures
Returns boolean vector of rejected hypotheses}
\usage{
rejected_hypotheses(object, ...)
}
\arguments{
\item{object}{Multiple testing procedure object (e.g. ddhwResult)}

\item{...}{Arguments passed to individual methods}
}
\value{
boolean vector of rejected hypotheses
}
\description{
Generic function for multiple testing procedures
Returns boolean vector of rejected hypotheses
}

