% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw_class.R
\name{rejections.ddhw}
\alias{rejections.ddhw}
\title{Get number of rejections by DDHW}
\usage{
rejections.ddhw(object, method = "adjpvals", groupwise = F)
}
\arguments{
\item{groupwise}{If F returns an integer (total number of rejections), if T return rejections within each stratum}
}
\description{
Get number of rejections by DDHW
}

