% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw_class.R
\name{thresholds.ddhw}
\alias{thresholds.ddhw}
\title{ddhw Accessor function for ddhw thresholds}
\usage{
thresholds.ddhw(object, levels_only = T)
}
\arguments{
\item{levels_only}{If T return 1 threshold per stratum, otherwise return a vector of thresholds of the same length as the original vector of p--values}
}
\description{
ddhw Accessor function for ddhw thresholds
}

