% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddhw_class.R
\name{weights.ddhw}
\alias{weights.ddhw}
\title{weights: Accessor function for ddhw weights}
\usage{
weights.ddhw(object, levels_only = T)
}
\arguments{
\item{levels_only}{If T return 1 weights per stratum, otherwise return a vector of weights of the same length as the original vector of p--values}
}
\description{
weights: Accessor function for ddhw weights
}

