% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ddhw_grouped}
\alias{ddhw}
\alias{ddhw_grouped}
\title{ddhw: Data-Driven Hypothesis Weights}
\usage{
ddhw_grouped(unadj_p, groups, alpha, mtests = length(unadj_p),
  optim_method = "MILP", solver = "Rsymphony", regularization_term = 0,
  optim_pval_threshold = "auto", time_limit = Inf)

ddhw(unadj_p, filter_statistic, nbins, alpha, ...)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Vector which assigns a group to each p-value.}

\item{alpha}{Numeric, sets the nominal level for FDR control.}

\item{mtests}{Numeric, true number of comparisons, must be at least ‘length(unadj_p)}

\item{optim_method}{Optimization method with which optimal weights are determined. Available options are "MILP" (default) and "subplex".}

\item{solver}{Solver to be used for "MILP" optimization. Available options are "Rsymphony" (default, open source) and "gurobi" (commercial).}

\item{regularization_term}{Numeric, biases weights to have the same magnitude, only available for optim_method=="subplex".}

\item{optim_pval_threshold}{Numeric between 0 and 1 or Character "auto" (default). P-values above this threshold get excluded
from the optimization procedure (e.g. because it is known a-priori that P-values above that threshold will
not get rejected), thus making it faster.}

\item{time_limit}{Numeric, sets maximum time limit for MILP solvers. Defaults to Inf.}

\item{filter_statistic}{Vector which contains the filter-statistic (covariate, independent under the H0) for each p-value}

\item{nbins}{Integer, number of groups into which p-values will be split based on filter_statistic.}
}
\description{
ddhw: Data-Driven Hypothesis Weights
}

